BASW-324: Fix Auto renewals on Phase 3 Workstream
Several scenarios for auto-renewal were faililng, for example, for payment
plans with a single installment, plans with only new memberships and plans
with only donations.

Refactored the query being used to obtain recurring contributions to renew
into separate methods to obtain single installment plans vs multiple
installment payment plans. Also rewrote queries to look for all scenarios,
based on line items associated to the recurring contribution set to auto-renew
and the specifics for each.

Finally, refactored the renewal of payment plans into two new classes, one
that handles single installment payment plans and the other one for plans with
multiple installments.
