+++ CRM/MembershipExtras/Job/OfflineAutoRenewal/SingleInstallmentPlan.php
@@ -1,0 +1,196 @@
+<?php
+use CRM_MembershipExtras_Service_MoneyUtilities as MoneyUtilities;
+use CRM_MembershipExtras_Service_InstallmentReceiveDateCalculator as InstallmentReceiveDateCalculator;
+
+/**
+ * Renews the payment plan and the related memberships if it paid by once and
+ * not using installments.
+ *
+ * Paid by once (no installments) payment plan get renewed by creating single
+ * pending contribution that links to the already existing recurring
+ * contribution.
+ */
+class CRM_MembershipExtras_Job_OfflineAutoRenewal_SingleInstallmentPlan extends CRM_MembershipExtras_Job_OfflineAutoRenewal_PaymentPlan {
+
+  /**
+   * @inheritdoc
+   */
+  public function renew() {
+    $this->duplicateRecurringLineItems($this->currentRecurContributionID);
+    $this->updateRecurringContributionAmount($this->currentRecurContributionID);
+    $this->setNewRecurringContribution();
+    $this->buildLineItemsParams();
+    $this->setTotalAndTaxAmount();
+    $this->paymentPlanStartDate = $this->calculateNoInstallmentsPaymentPlanStartDate();
+
+    $this->recordPaymentPlanFirstContribution();
+    $this->renewPaymentPlanMemberships();
+  }
+
+  /**
+   * Sets new recurring contribution from current recurring contribution.
+   */
+  private function setNewRecurringContribution() {
+    $this->newRecurringContributionID = $this->currentRecurContributionID;
+    $this->newRecurringContribution = civicrm_api3('ContributionRecur', 'getsingle', [
+      'id' => $this->newRecurringContributionID,
+    ]);
+  }
+
+  /**
+   * Creates duplicates of line items associated to recurring contribution that
+   * are set to auto-renew, don't have an end date and are not removed. It sets
+   * end date for the old line items.
+   *
+   * @param $recurringContributionID
+   */
+  private function duplicateRecurringLineItems($recurringContributionID) {
+    $lineItems = civicrm_api3('ContributionRecurLineItem', 'get', [
+      'sequential' => 1,
+      'contribution_recur_id' => $recurringContributionID,
+      'auto_renew' => 1,
+      'is_removed' => 0,
+      'end_date' => ['IS NULL' => 1],
+      'api.LineItem.getsingle' => [
+        'id' => '$value.line_item_id',
+        'entity_table' => ['IS NOT NULL' => 1],
+        'entity_id' => ['IS NOT NULL' => 1]
+      ],
+      'options' => ['limit' => 0],
+    ]);
+
+    foreach ($lineItems['values'] as $line) {
+      $lineItemParams = $line['api.LineItem.getsingle'];
+      unset($lineItemParams['id']);
+      $lineItemParams['unit_price'] = $this->calculateLineItemUnitPrice($lineItemParams);
+      $lineItemParams['line_total'] = MoneyUtilities::roundToCurrencyPrecision($lineItemParams['unit_price'] * $lineItemParams['qty']);
+      $lineItemParams['tax_amount'] = $this->calculateLineItemTaxAmount($lineItemParams['line_total'], $lineItemParams['financial_type_id']);
+
+      $newLineItem = civicrm_api3('LineItem', 'create', $lineItemParams);
+
+      $newStartDate = $this->calculateNoInstallmentsPaymentPlanStartDate();
+      CRM_MembershipExtras_BAO_ContributionRecurLineItem::create([
+        'contribution_recur_id' => $recurringContributionID,
+        'line_item_id' => $newLineItem['id'],
+        'start_date' => $newStartDate,
+        'auto_renew' => 1,
+      ]);
+
+      $endDate = new DateTime($newStartDate);
+      $endDate->sub(new DateInterval('P1D'));
+      CRM_MembershipExtras_BAO_ContributionRecurLineItem::create([
+        'id' => $line['id'],
+        'end_date' => $endDate->format('Y-m-d'),
+      ]);
+    }
+  }
+
+  /**
+   * Calculates the new start date for the payment plan
+   * if its paid with no installments.
+   * @return string
+   */
+  private function calculateNoInstallmentsPaymentPlanStartDate() {
+    $currentRecurContribution = civicrm_api3('ContributionRecur', 'get', [
+      'sequential' => 1,
+      'id' => $this->currentRecurContributionID,
+    ])['values'][0];
+    $installmentReceiveDateCalculator = new InstallmentReceiveDateCalculator($currentRecurContribution);
+
+    $paymentPlanContributionsCount = civicrm_api3('Contribution', 'getcount', [
+      'contribution_recur_id' => $this->currentRecurContributionID,
+    ]);
+
+    return $installmentReceiveDateCalculator->calculate($paymentPlanContributionsCount + 1);
+  }
+
+  /**
+   * @inheritdoc
+   */
+  protected function getRecurringContributionLineItemsToBeRenewed($recurringContributionID) {
+    $lineItems = civicrm_api3('ContributionRecurLineItem', 'get', [
+      'sequential' => 1,
+      'contribution_recur_id' => $recurringContributionID,
+      'auto_renew' => 1,
+      'is_removed' => 0,
+      'end_date' => ['IS NULL' => 1],
+      'api.LineItem.getsingle' => [
+        'id' => '$value.line_item_id',
+        'entity_table' => ['IS NOT NULL' => 1],
+        'entity_id' => ['IS NOT NULL' => 1],
+      ],
+    ]);
+
+    if (!$lineItems['count']) {
+      return [];
+    }
+
+    $result = [];
+    foreach ($lineItems['values'] as $line) {
+      $lineData = $line['api.LineItem.getsingle'];
+      $result[] =  $lineData;
+    }
+
+    return $result;
+  }
+
+  /**
+   * @inheritdoc
+   */
+  protected function getNewPaymentPlanActiveLineItems() {
+    $lineItems = civicrm_api3('ContributionRecurLineItem', 'get', [
+      'sequential' => 1,
+      'contribution_recur_id' => $this->newRecurringContributionID,
+      'is_removed' => 0,
+      'end_date' => ['IS NULL' => 1],
+      'api.LineItem.getsingle' => [
+        'id' => '$value.line_item_id',
+        'entity_table' => ['IS NOT NULL' => 1],
+        'entity_id' => ['IS NOT NULL' => 1],
+      ],
+    ]);
+
+    if (!$lineItems['count']) {
+      return [];
+    }
+
+    $result = [];
+    foreach ($lineItems['values'] as $line) {
+      $lineData = $line['api.LineItem.getsingle'];
+      $result[] =  $lineData;
+    }
+
+    return $result;
+  }
+
+  /**
+   * @inheritdoc
+   */
+  protected function calculateRecurringContributionTotalAmount($recurringContributionID) {
+    $totalAmount = 0;
+
+    $result = civicrm_api3('ContributionRecurLineItem', 'get', [
+      'sequential' => 1,
+      'contribution_recur_id' => $recurringContributionID,
+      'start_date' => ['IS NOT NULL' => 1],
+      'is_removed' => 0,
+      'end_date' => ['IS NULL' => 1],
+      'api.LineItem.getsingle' => [
+        'id' => '$value.line_item_id',
+        'entity_table' => ['IS NOT NULL' => 1],
+        'entity_id' => ['IS NOT NULL' => 1],
+      ],
+      'options' => ['limit' => 0],
+    ]);
+
+    if ($result['count'] > 0) {
+      foreach ($result['values'] as $lineItemData) {
+        $totalAmount += $lineItemData['api.LineItem.getsingle']['line_total'];
+        $totalAmount += $lineItemData['api.LineItem.getsingle']['tax_amount'];
+      }
+    }
+
+    return MoneyUtilities::roundToCurrencyPrecision($totalAmount);
+  }
+
+}
